All ground functions return now a double value. Improved newton iteration. The derivation decides whether the moving direction is positiv|negativ the returning value is negativ|positiv. This fixed the most common bugs in the ground-object collision except the slope of the linear function is between 0.15 and -0.15. In this case the collision is rather bad (impossible). Furthermore the newton algorithm decides about the deep of the iteration. Higher object speed causes a more longer calculation. That will be replaced by a more intelligent algorithm soon.

+++ PhysicsEngine/src/de/engine/objects/Ground.java
@@ -40,7 +40,7 @@ public class Ground //extends ObjectProperties
         // what height are you going to go?
         int hill_height = 100;
         
-        return (int) (Math.sin(phase * i * Math.PI / 300) * hill_height + hill_height) + this.watermark;
+        return Math.sin(phase * i * Math.PI / 300) * hill_height + hill_height + this.watermark;
     }
     
     
@@ -54,9 +54,9 @@ public class Ground //extends ObjectProperties
         // what height are you going to go?
         int hill_height = 100;
         // Grass heigth
-        int grass_heigth = 4;
+        int grass_heigth = 4; 
         
-        return (int) (Math.sin(phase * i * Math.PI / 360) * hill_height + hill_height + grass_heigth*Math.sin(i)) + this.watermark;
+        return Math.sin(phase * i * Math.PI / 360) * hill_height + hill_height + grass_heigth*Math.sin(i) + this.watermark;
     }
     
     
@@ -68,7 +68,7 @@ public class Ground //extends ObjectProperties
         // stairs 
         int stairs = 12;
         
-        return (int) (-0.3*i - (9*i*i + stairs*i*i*i - i)/(i-150*i*i + 200) + 50 + 4d*Math.sin(0.04*i)) + this.watermark;
+        return -0.3*i - (9*i*i + stairs*i*i*i - i)/(i-150*i*i + 200) + 50 + 4d*Math.sin(0.04*i) + this.watermark;
     }
